requirements.txt
python-dotenv
django
django-rest_framework
django-cors-headers
djangorestframework-simplejwt
requests
Pillow
celery
django_celery_beat
channels
channels_redis
daphne
uvicorn
websockets
wsproto
psycopg2-binary
psycopg
psycopg2
